WindowServer+LibGfx: Move CursorParams to LibGfx
They will be used by MouseSettings in the next commit.
